refs pace-neutrons/Herbert#14 a bit more tests and better tracking for set empty array
git-svn-id: https://svn.isis.rl.ac.uk/Herbert/trunk@327 6b2803e2-fb1c-43d7-acfb-f5118c1e7254

+++ classes/data_loaders/@a_loader/private/set_consistent_array.m
@@ -8,7 +8,12 @@ function this = set_consistent_array(this,field_name,value)
 %
 
 if isempty(value)
-    this=this.delete();
+    if isempty(this.file_name)
+        this=this.delete();
+    else
+        this.S_stor=[];
+        this.ERR_stor=[];
+    end
     return
 end
